+++ catalog/app/utils/BucketPreferences/BucketPreferences.spec.ts
@@ -12,7 +12,9 @@ const expectedDefaults = {
     },
     blocks: {
       analytics: true,
-      browser: true,
+      browser: {
+        hidden: true,
+      },
       code: true,
       meta: {
         userMeta: {