+++ catalog/app/utils/BucketPreferences/index.ts
@@ -1,10 +1,2 @@
-export type {
-  ActionPreferences,
-  AthenaPreferences,
-  MetaBlockPreferences,
-  NavPreferences,
-  PackagePreferences,
-  SourceBuckets,
-} from './BucketPreferences'
-export { Result } from './BucketPreferences'
+export * from './BucketPreferences'
 export * from './Provider'